improve naming + improve code.
